Adding prob_end to match class
This is with a view to addressing 37. This is just a small step but
going in the correct direction.

- Matches can be given a `prob_end` argument which is the probability
  that a match ends on a given round.
- This is randomly sampled using inverse sampling so that we're not
  querying the need for it and the random module more than once.
- At present this can be used 'in conjunction' with `turns` so that a
  match stops either before some number of turns or probabilistically.
- Docs have been updated.

Note that this does not break any other functionality, not passing a
prob_end means the match just behaves as normal.

I have also taken the opportunity to add hypothesis tests to the match
class.