Reverting 461
This is done with a view to clearing the way for #37 which will be made
far easier with #493.

The plan being that if a match sets the information available to players
then a probabilistic ending is just equivalent to randomly sampling the
length of the match (as I had done in the PR that this reverts) and not
letting the players know.