Re write of run_axelrod.
This does not use the Manager or ManagerFactory class, no doubt some
duplication can be reduced but perhaps this is an OK balance?

The only functionality lost from the current run_axelrod (I believe) is
the logger. I'm not entirely sure I understand the need for it (so
please correct me, perhaps the logger object itself can be lifted in to
here if it's necessary?).