Altering the arguments of super function in  Spatial Match Generator.
SpatialMatches is class created in the match generator file in order to create matches where not all players play each other.
But instead the players, that are illustrated as nodes in a graph, only play with players that are connected to by an edge.
SpatialMatches inherit from RoundRobinMatches but there is need for an extra argument taken by the class which is an edge
dictionary. To achieve that we use the python function super.

Super takes the name of the class and a self argument, and
here we will replace what was previously written as super(RoundRobinMatches, self) with super(SpatialMatches, self).