Tests for the class SpatialTournament in the Tournament file.
SpatialTournament is a class which will be running a Spatial Tournament by calling the SpatialMatches class
we created before. It inherit for the class Tournament and these are some tests for the specific class.

The tests for the SpatialTournament do not contain the hypothesis tests yet. They will be added shortly.
We have run all the unit tests and we all 22 have passed.