Adding Spatial Tournament Class.
The Spatial Tournament class is a class resposinble for creating a Spatial structure tournament, where the players
are respresented on a graph and they only matches are bewteen players-nodes connected with an edge. It does that
by using the SpatialMatch class and it inherit from the tournament class.

It's take a extra argument which is a dictionary called edges, that represents which nodes- players are connected
and will face each other in the tournament.

Proper testes for this class have been written and all are passing.