Adding a Value Error for tournaments with disconnected nodes-players.
In case a tournament is created containing disconnected nodes-players an Error will be displayed.
That is because if a player does not have any interaction with any other player it should be removed from the tournament
as not active player  could not have a score or rank in the tournament. A player not participating is different from a player
scoring zero and this should be noted.

Also a test for this constrain was added in the test_match_generator.