Add a better doctest for setting random seed.
The previous doctests were checking particular results. When new
strategies where added to the library and/or other changes were made,
this could offset the random seed so the exact result would not
necessarily be the same.

This is the actual check we want: it checks that the results don't
change. We don't care what they actually are.