Add version number to library
Closes #731.

We cannot import __version__ from __init__.py in setup.py. This would
load all of axelrod which at that point does not have any dependencies
installed. The approach here is suggested in the later answers here:
http://stackoverflow.com/questions/458550/standard-way-to-embed-version-into-python-package

Have written a test: this means that when bumping the number we will
need to bump it in two places (`version.py` and `test_version.py`). If
we think this is OK I will update the wiki for the release process to
include: "run the tests" which probably should be there anyway.