Use the random module for sampling team.
Using numpy causes problems for building the docs. Building the docs
mocks numpy (I believe this is because readthedocs can't handle numpy),
which in turn broke when importing axelrod (for the auto documentation)
as various filtered lists are created when importing. MWE Random is the
first strategy that initialises and classifies itself stochasticly...