Closes #776
Passing matplotlib axes.

Giving access to the matplotlib axes of the plots
in the library. If user does not specify the axes
then it uses `plt.subplots`.

Some measures, such as number of players, players
etc were used in more than one plot so I initiliaze
them in the class. Also, I have not yet removed the
title argument because I am not sure if is needed
for when running our tournament and saving the plots.

Please let me know you suggestions.