Add state to action analysis.
Add two functions to interaction_utils:

- `interaction_utils.compute_state_to_action_distribution`: computes
  list of counter objects of mapping a `(state, action)` to a count.
- `interaction_utils.compute_normalised_state_to_action_distribution`: computes
  list of counter objects of mapping a `(state, action)` to a normalised
  count (for a given state).

There have then been incorporated in `axelrod.result_set` and in particular in
the `result_set.summarise()` output. Which now includes the average
rates for each strategy in a tournament. So for memory 1 strategies this
means we can measure their parameters and for non memory 1 strategies we
can infer what the parameters correspond to the data.