Fix bug in base reset case.
We were not actually testing anything here. We were comparing the value
of the clone's attribute to the value of the clone's attribute.

I have fixed this (and docs). Note I also changed the test to iterate
over the player so that we can see if any new attributes have been
picked up along the way.