Refactor averagecopier for 884 and a bug.
Note that I think I've caught a bug here:

The average copier was using integer division for the calculation of the
opponent's cooperation ratio. So it was in effect either 0 or 1. I've
fixed this and also written tests that show the behaviour changes for
different seeds (in essence testing that the behaviour is stochastic).