Update test_worse_and_worse.py
I deleted the last test for the class worse and worse 3, the test before it test the given seed for a 0/1 probability of defecting. The deleted was giving the same result as the prior Defector test . I went ahead and fixed the other corrections and ran the unitteest on each class, all checked out. HopefullyI am getting all the kinks out now, and my next PR's hopefully are a lot better.