Ensure that transformers do not overwrite docstrings
Currently our docstrings get over written for a number of strategies.
For example the Alexei strategy's docstring does not appear (as the
class written passes through a transformer):
http://axelrod.readthedocs.io/en/latest/reference/all_strategies.html#axelrod.strategies.titfortat.Alexei

(Here is the source code for Alexei:
http://axelrod.readthedocs.io/en/latest/_modules/axelrod/strategies/titfortat.html#Alexei

This ensures the original docstring is kept and also fixes a couple of
errors that then appeared in the docstrings.