Correct SteinRap strategy and add relevant test.
The strategy wasn't quite correct: it would in fact not
remember the result of the chisquared test from one call to another.

I've fixed that here. I've also included a few more tests including a
test that the chi squared test gets run again and that it can change
it's mind.