Make match decide on turns in match.play
All matches have both a turns and a prob_end attribute.

Standard matches have a finite turns and a prob_end of 0.
Prob end matches have an infinite number of turns and a prob_end > 0.

The play method of the match is where a Match "decides" how many turns
to play for. This ensures that the cache will work properly.

This also implies that the len of an infinite match is simply not
defined and raises an error.

Also changed how the length of a prob end match is sampled: do not
sample the length if not required.