Remove morality_metrics as arg for Tournament.
On https://github.com/Axelrod-Python/Axelrod/pull/1042 @marcharper
mentioned that it was perhaps worth moving this argument out of the
Tournament.init.

I took a look and turns out that this argument is not being used anymore
(I'm not sure since when). Probably since the move to doing the result
calculations on disk.