Remove hypothesis from setup.
Currently when we pip install axelrod we also install hypothesis.

This is however only used for development. I've added a tweak to the
setup.py file to just remove it from the read in requirements.

I don't think we need to adjust the contributing documentation as `pip
install -r requirements` will still install hypothesis.