Implement transitive fingerprint (#1125)
* Implement transitive fingerprint

Transitive fingerprints give us another fingerprint
for the library. Is showing the cooperation rate
of a strategy against a set of opponents over the turns.

The default opponents are a list of random player with
increasing cooperation probability.

This is an implementation of fingerprints
similar to the ones used in:

- https://arxiv.org/abs/1707.06920
- https://arxiv.org/abs/1707.06307

- write tests for transitive fingerprint
- Write docstrings.
- Add documentation for transitive fingerprints
- Remove the optional matplotlib check: This was already removed elsewhere with bc952f4

* corrected typos

* fix fingerprints to fingerprint

* Add better ylabel for defaults.

* Update plot for docs.

* Add ability to pass an axes to the plot. (#7)

This means you can more easily change font size etc.

* Implement transitive fingerprint

Transitive fingerprints give us another fingerprint
for the library. Is showing the cooperation rate
of a strategy against a set of opponents over the turns.

The default opponents are a list of random player with
increasing cooperation probability.

This is an implementation of fingerprints
similar to the ones used in:

- https://arxiv.org/abs/1707.06920
- https://arxiv.org/abs/1707.06307

- write tests for transitive fingerprint
- Write docstrings.
- Add documentation for transitive fingerprints
- Remove the optional matplotlib check: This was already removed elsewhere with bc952f4

* corrected typos

* fix fingerprints to fingerprint

* Add better ylabel for defaults.

* Update plot for docs.

* fix typos pointed by Marc

* replace docstrings full word to py function

* rename number_opponents/ number_of_opponents

* Change default repetitions to 1000.

* Run smaller image to correct repetitions.