add implementation of a generic memory two + 2 strategies
Close #1063

A generic class for memory two players, similar to the MemoryOnePlayer class.
If transition rates are not given the the player is set to default which
is Cooperator.

I have added the equivilent tests and a reference from issue #1063.

Moreover, this implements two strategies of interest from #1063:

- AON2 and Delayed AON1

Each strategy has 16 states. I have added tests for most of the states.