fixes initialisation bug for mem one
This closes issue #1169.

The issue has been that if `four_vector` was not set the player would not
break. This is fixed by having a default `four_vector` which corresponds
to the strategy Win-Stay Lose-Shift. Note that when this happens a
warning message is raised to let the user know (using the library
warnings).

A test has been altered to test that the default values are passed correctly.
Test: `test_default_if_four_vector_not_set`.

An issue has been that the players that inherited from `MemoryOnePlayer`
were breaking. This was because several classes were initialising with
an empty vector and were setting the vector are retrieving information
on the game.

This was addressed using the function `get_initial_four_vector`, which
will only set a default and raise an error for the general case and
will ignore the fact that `four_vector` is empty for the classes which
have been breaking.