add implementation of a generic memory two + 2 strategies (#1171)
* add implementation of a generic memory two + 2 strategies

Close #1063

A generic class for memory two players, similar to the MemoryOnePlayer class.
If transition rates are not given the the player is set to default which
is Cooperator.

I have added the equivilent tests and a reference from issue #1063.

Moreover, this implements two strategies of interest from #1063:

- AON2 and Delayed AON1

Each strategy has 16 states. I have added tests for most of the states.

* add AON2 and Delayed AON1 to all strategies

* add to the documentation of the strategies

Add information about the vectors which are reported as equivilent.
I am not sure I can say that they are actually different
in the documentation without saying why and without adding the
fingerprints.

* Fix type hint. (#8)

* remove initial action option

* minor tweak to documentation

* add to tests initialisation of stochastic mem two

* fix doc test

* test stochastic behaviour of memory two player

Added tests against two opponents. For different seed
the behaviour of the memory two strategy changes.

* address comments on pr by Marc