WIP: Fix compute features and retrain ann (#1182)
* Fix weight parametrization for `my_previous_d`

Corresponds to #1172

* Add new weights.

Note that I trained ANNNoise with a hidden layer of size 10 due to the
docstring. This was in fact an incorrect (the layer had size 5)
docstring but retraining took
quite a while...

* Update tests.

* Remove depreciated dask call.

* Add link to archived training data.

* Sep. archived training data and fix name.

This fails tests, waiting on new data.

* Update training weight for noisy ANN.

* Remove unneeded autodoc calls

This is causing a failure with latest version of sphinx. The docs still
build fine.