Modify/fix initial play of lookerup players.
Closes #1185

Currently, the lookerup players pop from a list of initial plays. This
breaks when playing against the cheating players which simulate their
play (as the list of initial players no longer exists).

This fixes that by ensuring the list of initial plays does not
disappear which I think technically is probably better behaviour anyway?