Re-implement Downing.
This is a complete rewrite of the Downing strategy.

To be able to do this I've used the description in Downing's 1975 paper.
This description itself is not sufficiently clear and so I've had to
make some further assumptions which I've clearly documented.

Note: there was documentation claiming that there was a bug in the
implementation in the original tournament. I believe this was a mistake
due to a misinterpretation of one online set of slides where they
commented that there was a mistake in the implementation. This however
was not a bug and was actually described quite a lot in Axelrod's
original tournament: the strategy was implemented to act a particular
way in the first two rounds and this had the result of making the
strategy a king maker. This however was not a bug, just a particular
interpretation of the overall decision rule described in Downing's 1975
paper.