Modify Shubik.
Note that this was not the specific error that @id428 pointed out but
having reviewed the papers and source code I found this one minor
inaccuracy.

@id428 made a point that the strategy should cooperate twice after it's
round of retaliations but I do
not see this in any of the descriptions of the strategy. Once the
strategy has finished retaliating, all the texts indicate that it
cooperates again but ready to retaliate.