Created CITATION.cff
I used cff init website to create a CITATION.cff file based on the author's information. There were some unclear given or family names. I made judgment calls, but please review them to ensure that they are correctly attributed. I included the readthedocs as a landing website and the GitHub's webpage as the location of the repository. Additional information about the authors (i.e., email, affiliation and ORCID) can optionally be added. I think the file's format is pretty clear but notify if that is not the case.